refactor: add app and nitro log - resolve dependencies issue
