+++ core/src/node/extension/extension.ts
@@ -1,7 +1,5 @@
 import { rmdirSync } from 'fs'
 import { resolve, join } from 'path'
-import { manifest, extract } from 'pacote'
-import * as Arborist from '@npmcli/arborist'
 import { ExtensionManager } from './manager'
 
 /**
@@ -41,6 +39,7 @@ export default class Extension {
    * @param {Object} [options] Options provided to pacote when fetching the manifest.
    */
   constructor(origin?: string, options = {}) {
+    const Arborist = require('@npmcli/arborist')
     const defaultOpts = {
       version: false,
       fullMetadata: false,
@@ -74,13 +73,15 @@ export default class Extension {
   async getManifest() {
     // Get the package's manifest (package.json object)
     try {
-      const mnf = await manifest(this.specifier, this.installOptions)
-
-      // set the Package properties based on the it's manifest
-      this.name = mnf.name
-      this.version = mnf.version
-      this.main = mnf.main
-      this.description = mnf.description
+      await import('pacote').then((pacote) => {
+        return pacote.manifest(this.specifier, this.installOptions).then((mnf) => {
+          // set the Package properties based on the it's manifest
+          this.name = mnf.name
+          this.version = mnf.version
+          this.main = mnf.main
+          this.description = mnf.description
+        })
+      })
     } catch (error) {
       throw new Error(`Package ${this.origin} does not contain a valid manifest: ${error}`)
     }
@@ -99,7 +100,8 @@ export default class Extension {
       await this.getManifest()
 
       // Install the package in a child folder of the given folder
-      await extract(
+      const pacote = await import('pacote')
+      await pacote.extract(
         this.specifier,
         join(ExtensionManager.instance.extensionsPath ?? '', this.name ?? ''),
         this.installOptions,
@@ -164,10 +166,13 @@ export default class Extension {
    * @returns the latest available version if a new version is available or false if not.
    */
   async isUpdateAvailable() {
-    if (this.origin) {
-      const mnf = await manifest(this.origin)
-      return mnf.version !== this.version ? mnf.version : false
-    }
+      return import('pacote').then((pacote) => {
+        if (this.origin) {
+        return pacote.manifest(this.origin).then((mnf) => { 
+          return mnf.version !== this.version ? mnf.version : false
+        })
+      }
+    })
   }
 
   /**
