+++ core/src/node/extension/manager.ts
@@ -1,7 +1,6 @@
 import { join, resolve } from "path";
 
 import { existsSync, mkdirSync, writeFileSync } from "fs";
-import { init } from "./index";
 import { homedir } from "os"
 /**
  * Manages extension installation and migration.
@@ -18,22 +17,6 @@ export class ExtensionManager {
     if (ExtensionManager.instance) {
       return ExtensionManager.instance;
     }
-  }
-
-  /**
-   * Sets up the extensions by initializing the `extensions` module with the `confirmInstall` and `extensionsPath` options.
-   * The `confirmInstall` function always returns `true` to allow extension installation.
-   * The `extensionsPath` option specifies the path to install extensions to.
-   */
-  setupExtensions() {
-    init({
-      // Function to check from the main process that user wants to install a extension
-      confirmInstall: async (_extensions: string[]) => {
-        return true;
-      },
-      // Path to install extension to
-      extensionsPath: join(userSpacePath, "extensions"),
-    });
   }
 
   setExtensionsPath(extPath: string) {
