+++ core/src/node/log.ts
@@ -3,16 +3,18 @@ import util from 'util'
 import path from 'path'
 import os from 'os'
 
-const appDir = path.join(os.homedir(), 'jan')
+export const logDir = path.join(os.homedir(), 'jan', 'logs')
 
-export const logPath = path.join(appDir, 'app.log')
-
-export const log = function (d: any) {
-  if (fs.existsSync(appDir)) {
-    var log_file = fs.createWriteStream(logPath, {
+export const log = function (message: string, fileName: string = 'app.log') {
+  if(!fs.existsSync(logDir)) {  
+    fs.mkdirSync(logDir, { recursive : true })
+  }
+  if (fs.existsSync(logDir)) {
+    var log_file = fs.createWriteStream(path.join(logDir, fileName), {
       flags: 'a',
     })
-    log_file.write(util.format(d) + '\n')
+    log_file.write(util.format(message) + '\n')
     log_file.close()
+    console.debug(message)
   }
 }
