+++ extensions/inference-nitro-extension/src/index.ts
@@ -21,6 +21,7 @@ import {
   Model,
   joinPath,
   InferenceExtension,
+  log,
 } from "@janhq/core";
 import { requestInference } from "./helpers/sse";
 import { ulid } from "ulid";
@@ -270,6 +271,7 @@ export default class JanInferenceNitroExtension implements InferenceExtension {
         }
         message.status = MessageStatus.Error;
         events.emit(EventName.OnMessageUpdate, message);
+        log(`[APP]::Error: ${err.message}`);
       },
     });
   }
